+++ src/vsuni.h
@@ -2,7 +2,7 @@
 #define _FCEU_VSUNI_H
 
 void FCEU_VSUniPower(void);
-void FCEU_VSUniCheck(uint64 md5partial, int *, uint8 *);
+void FCEU_VSUniCheck(uint64 md5partial, int *, int *);
 void FCEU_VSUniDraw(uint8 *XBuf);
 
 void FCEU_VSUniToggleDIP(int);	/* For movies and netplay */
